Update problem5.py
This is not my code, I only added the comments. Just thought it was the best code solution for problem5-Project Euler. Do not want to take credit,but this is the type of code I would like to write someday. I decided to keep it in this repository for further reference